fixed system process to have errors returned, task runner refactor

+++ lib/cassie/tasks/migration/import.rake
@@ -31,7 +31,7 @@ namespace :cassie do
         importer.import
         puts "-- done"
       rescue => e
-        importer.imported_paths.each {|f| File.delete(f) }
+        importer.imported_paths.each {|f| File.delete(f) } if importer
         output_error(e)
         abort
       end
